[#47] Fix code style and docs.
