+++ legion/bin/liara
@@ -25,7 +25,7 @@ import os
 import argparse
 import logging
 
-from legion.liara.server import liara
+from legion.liara.server import LIARA
 import legion.utils
 
 ROOT_LOGGER = logging.getLogger()
@@ -57,4 +57,4 @@ if __name__ == '__main__':
 
     ROOT_LOGGER.setLevel(log_level)
 
-    liara.start(args)
+    LIARA.start(args)
